add another type of loop optimizable, LocalRoutineCall
